# Reduce compilation time by 75%, memory usage by 63%