@jcarpent 2020-03-28
@wxmerkt This weird, because this does not seem to be linked together, as the function you moved is a template function, never called by any program (like Pinocchio or Croco).

Could give the commands to measure performances?