@wxmerkt 2020-03-28
This was measured solely within EigenPy and the reductions are w.r.t. EigenPy. You can measure (with ccache deactivated, cache deleted, etc.): `/usr/bin/time -v make -j1`. I reported the benchmarks above from clean compilations on the same machine.

I have not yet done similar effort in Pinocchio or Crocoddyl - but the underlying symptom of spikes in memory and compilation time usage are similar.