@jcarpent 2020-03-28
I will give a try on my side and let you know. For Windows, the main issue is that the symbols are not exported.