@jcarpent 2020-03-28
Here a my results:

# Before
---------
```
	Command being timed: "make -j1"
	User time (seconds): 348.34
	System time (seconds): 12.74
	Percent of CPU this job got: 95%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 6:17.16
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 1376592
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 8544
	Minor (reclaiming a frame) page faults: 4186873
	Voluntary context switches: 5580
	Involuntary context switches: 510369
	Swaps: 0
	File system inputs: 0
	File system outputs: 0
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 106
	Page size (bytes): 4096
	Exit status: 0
```
# After
-----
```
	Command being timed: "make -j1"
	User time (seconds): 101.68
	System time (seconds): 5.99
	Percent of CPU this job got: 96%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 1:51.81
	Average shared text size (kbytes): 0
	Average unshared data size (kbytes): 0
	Average stack size (kbytes): 0
	Average total size (kbytes): 0
	Maximum resident set size (kbytes): 497832
	Average resident set size (kbytes): 0
	Major (requiring I/O) page faults: 1886
	Minor (reclaiming a frame) page faults: 1805334
	Voluntary context switches: 1831
	Involuntary context switches: 109283
	Swaps: 0
	File system inputs: 0
	File system outputs: 0
	Socket messages sent: 0
	Socket messages received: 0
	Signals delivered: 106
	Page size (bytes): 4096
	Exit status: 0
```
Your suggestion seems indeed to reduce a lot the compilation timings