@jcarpent 2020-03-28
In fact, your suggestion is not compiling at the exposition of classic types. By moving the template to the .cpp file, the symbols `void eigenpy::exposeType<std::__1::complex<double>, 0>()` are not generated at all. This is why the compilation has decreased so much.