@wxmerkt 2020-03-28
Hmmm, the unit tests passed though - this is weird? Either way, I am very happy for any suggestions that get memory and compile time down or fixes to the above solution.