@jcarpent 2020-03-28
It is maybe because you've another libeigenpy.so located in one of your LD_LIBRARY_PATH ;)