@wxmerkt 2020-03-28
@jcarpent I actually did not remove the signatures from the header. It should work again after reverting your commit unless I am missing a point?