@jcarpent 2020-03-28
The thing is that the definition of the declared function does not appear any more. In other words, the code is no more visible and the templated function will appear as an empty symbol in the generated lib.