@wxmerkt 2020-03-28
Thanks, there was an artefact issue indeed making the unit tests appear to pass. I will see if there is another way to reduce compilation memory (we are at 3.x GB and should get below 2 GB).